Add storage pool support
LXD has grown support for multiple storage pools
defined by name and type (btrfs, zfs, lvm, etc)
and we would like to support this in nova-lxd

+++ nova/virt/lxd/flavor.py
@@ -98,19 +98,19 @@ def _root(instance, client, *_):
             specs['quota:disk_write_iops_sec'])
 
     if specs.get('quota:disk_read_bytes_sec'):
-        device['limits.read'] = '{}MB'.format(
-            int(specs['quota:disk_read_bytes_sec']) / units.Mi)
+        device['limits.read'] = '{}MB'.format(int(
+            int(specs['quota:disk_read_bytes_sec']) / units.Mi))
     if specs.get('quota:disk_write_bytes_sec'):
-        device['limits.write'] = '{}MB'.format(
-            int(specs['quota:disk_write_bytes_sec']) / units.Mi)
+        device['limits.write'] = '{}MB'.format(int(
+            int(specs['quota:disk_write_bytes_sec']) / units.Mi))
 
     minor_quota_defined = 'limits.write' in device or 'limits.read' in device
     if specs.get('quota:disk_total_iops_sec') and not minor_quota_defined:
         device['limits.max'] = '{}iops'.format(
             specs['quota:disk_total_iops_sec'])
     if specs.get('quota:disk_total_bytes_sec') and not minor_quota_defined:
-        device['limits.max'] = '{}MB'.format(
-            int(specs['quota:disk_total_bytes_sec']) / units.Mi)
+        device['limits.max'] = '{}MB'.format(int(
+            int(specs['quota:disk_total_bytes_sec']) / units.Mi))
     return {'root': device}
 
 
@@ -168,16 +168,16 @@ def _network(instance, _, network_info, __):
             int(specs.get('quota:vif_inbound_peak', 0)),
         )
         if vif_inbound_limit:
-            devices[key]['limits.ingress'] = '{}Mbit'.format(
-                vif_inbound_limit * units.k * 8 / units.M)
+            devices[key]['limits.ingress'] = '{}Mbit'.format(int(
+                vif_inbound_limit * units.k * 8 / units.M))
 
         vif_outbound_limit = max(
             int(specs.get('quota:vif_outbound_average', 0)),
             int(specs.get('quota:vif_outbound_peak', 0)),
         )
         if vif_outbound_limit:
-            devices[key]['limits.egress'] = '{}Mbit'.format(
-                vif_outbound_limit * units.k * 8 / units.M)
+            devices[key]['limits.egress'] = '{}Mbit'.format(int(
+                vif_outbound_limit * units.k * 8 / units.M))
     return devices
